DRY for perform_stream_db_test, #48
